+++ plugins/woocommerce-admin/client/header/index.js
@@ -21,7 +21,7 @@ import { Text, useSlot } from '@woocommerce/experimental';
 import './style.scss';
 import useIsScrolled from '../hooks/useIsScrolled';
 import { TasksReminderBar, useActiveSetupTasklist } from '../task-lists';
-import { useLaunchYourStore, LaunchYourStoreStatus } from './launch-your-store';
+import { LaunchYourStoreStatus, useLaunchYourStore } from './launch-your-store';
 
 export const PAGE_TITLE_FILTER = 'woocommerce_admin_header_page_title';
